Add dummy image cards
